Authorization Protocol Refactor
Refactor the method of providing authorization protocols to be handled
by standalone classes in order to simplify the Session class. Create
Basic and Cookie based authentication protocols for use by Session. Add
HTTP error response codes as exceptions to throw when accessing remote
resources and the server responds in an unexpected or undesirable
manner.